app/requirements.txt

requests
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-psycopg2
opentelemetry-exporter-otlp
prometheus-client